Added titration MPRA dataset from 06/31/17 from 0 - 64 µM Forsk
Added titration MPRA dataset from 06/31/17 and its processing. The  dataset has the same variants but at the higher concentrations of Forskolin: [0, 1, 2, 4, 8, 16, 25, 32, 64 µM]. Shared concentrations to the 09/21/17 dataset are 0, 1, and 4 µM.

Set-up 09/21/17 dataset to compare concentrations. Also created df of all expressions and concentrations such that concentration and expression is a single variable for each to make plotting response curves easier.

Will compare similar concentrations to look for correlation/same values for ease of adding expression to the response curves in 09/21/17 dataset.